Driver & TransactionReport updated fatory and test added

+++ src/main/java/domain/Driver.java
@@ -1,10 +1,8 @@
 package domain;
 
-import java.util.Date;
-
 /**
  * POJO Class Driver
- * Name: Gilberto Silva
+ * @author: Gilberto Silva
  * Student: 218239300
  */
 public class Driver {
@@ -13,7 +11,7 @@ public class Driver {
     private String last_Name;
     private int driver_License_Number;
     private String vehicle_Information;
-    private int phone_Number;
+    private long phone_Number;
     private String email;
 
     //-----------Constructors--------------
@@ -41,7 +39,7 @@ public class Driver {
 
     public String getLast_Name() {return last_Name;}
 
-    //------------Setters--------------
+
     public void setLast_Name(String last_Name) {this.last_Name = last_Name;}
 
     public int getDriver_License_Number() {return driver_License_Number;}
@@ -52,9 +50,9 @@ public class Driver {
 
     public void setVehicle_Information(String vehicle_Information) {this.vehicle_Information = vehicle_Information;}
 
-    public int getPhone_Number() {return phone_Number;}
+    public long getPhone_Number() {return phone_Number;}
 
-    public void setPhone_Number(int phone_Number) {this.phone_Number = phone_Number;}
+    public void setPhone_Number(long phone_Number) {this.phone_Number = phone_Number;}
 
     public String getEmail() {return email;}
 
@@ -79,7 +77,7 @@ public class Driver {
         private String last_Name;
         private int driver_License_Number;
         private String vehicle_Information;
-        private int phone_Number;
+        private long phone_Number;
         private String email;
 
         public Builder setDriver_ID(int driver_ID){
@@ -107,7 +105,7 @@ public class Driver {
             return this;
         }
 
-        public Builder setPhone_Number(int phone_Number){
+        public Builder setPhone_Number(long phone_Number){
             this.phone_Number = phone_Number;
             return this;
         }
